Update visitor to accept arbitrary template args.
Wrap the entirety of visitor in a template.
This allows separation of the compile-time args describing the
argument type (Args...) and other compile time args (such as lambdas).

The template returns a helper that looks similar to the old `visitor`
implementation but now has knowledge of this second set of compile-time
args that are captured by the surrounding template.

If there is at least one of these args, the mixin now includes a
"!(TemplateArgs)" that it sticks onto the member invocation.