Implement `pick` to choose between return types.
Pick wraps one of several values in a SuperStruct that encompasses all
the argument types. It is esentially a more general version of
std.range.choose/chooseAmong.o

TODO:
does it need a separate overload for bool/size_t (choose/chooseAmong)?
or is the implicit conversion from bool to size_t fine?