Wrap uncommmon return types in a SuperStruct.
If the sub-types of a SuperStruct all have a member that can be invoked
with the same arguments and return uncommon (but non-void) types, this
member _will_ be supported, and the return type will be a SuperStruct of
the possible return types.

If at least one member returns void, all other members must return void
as well to be supported.