Implement get!T to simulate casting to a subtype.
Allow casting a superstruct to a subtype by calling get!T.
This is provided as a function separate from opCast to avoid confusion
with builtin opCast forwarding.
It is provided as a free function, so users can avoid importing it if
it would conflict with a forwarded member named get on their subtypes.